feat(service&controller): JWT 기반 인증 시스템 구축 및 사용자 정보 리팩토링
- profile 관련 API Mapping 주소 변경
- 비밀번호 변경 시 현재 비밀번호로 인증하는 기능 추가
- 변경사항 반영해 관련 테스트 업데이트

+++ src/main/java/com/fourspoons/mikkureomi/user/service/UserService.java
@@ -4,12 +4,9 @@ package com.fourspoons.mikkureomi.user.service;
 import com.fourspoons.mikkureomi.constants.ErrorMessage;
 import com.fourspoons.mikkureomi.jwt.JwtTokenProvider;
 import com.fourspoons.mikkureomi.profile.service.ProfileService;
-import com.fourspoons.mikkureomi.user.dto.LoginRequestDto;
-import com.fourspoons.mikkureomi.user.dto.LoginResponseDto;
-import com.fourspoons.mikkureomi.user.dto.UpdatePasswordRequestDto;
+import com.fourspoons.mikkureomi.user.dto.*;
 import com.fourspoons.mikkureomi.user.repository.UserRepository;
 import com.fourspoons.mikkureomi.user.domain.User;
-import com.fourspoons.mikkureomi.user.dto.SignUpRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -80,9 +77,16 @@ public class UserService {
         user.updatePassword(passwordEncoder.encode(dto.getNewPassword()));
     }
 
-    public void deleteUser(Long userId) {
+    public void deleteUser(Long userId, PasswordRequestDto dto) {
         User user = userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
+                .orElseThrow(() -> new IllegalArgumentException(ErrorMessage.USER_NOT_FOUND.getMessage()));
+
+        String rawPassword = dto.getPassword();
+
+        if (!passwordEncoder.matches(rawPassword, user.getPassword())) {
+            throw new IllegalArgumentException(ErrorMessage.INVALID_PASSWORD.getMessage());
+        }
+
         profileService.deleteProfile(userId);
         userRepository.delete(user);
     }
